+++ packages/lib/env.d.ts
@@ -1,4 +1,4 @@
 export const env: {
-  mobxVersion: 4 | 5 | 6
+  mobxVersion: 4 | 5 | 6 | 7
   compiler: "tsc" | "tsc-experimental-decorators" | "babel" | "swc"
 }